[TS migration] Migrate 'ScreenWrapper' component to TypeScript
